requirements.txt
requests>=2.22.0
discord.py[voice]>=1.3.0
opuslib>=3.0.1
ffmpeg>=1.4
fuzzywuzzy[speedup]>=0.18.0
youtube-dl>=2020.5.3
